Moved place overview to a new file. Updated place overview styles
