+++ static/js/place/dev_place_main.tsx
@@ -1,4 +1,3 @@
-/** @jsxImportSource @emotion/react */
 /**
  * Copyright 2024 Google LLC
  *
@@ -14,33 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/** @jsxImportSource @emotion/react */
 
-import { DataRow } from "@datacommonsorg/client";
 import {
   PlaceChartsApiResponse,
   RelatedPlacesApiResponse,
 } from "@datacommonsorg/client/dist/data_commons_web_client_types";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { RawIntlProvider } from "react-intl";
 
-import { css, ThemeProvider, useTheme } from "@emotion/react";
-import { LocationCity } from "../components/elements/icons/location_city";
-import { GoogleMap } from "../components/google_map";
+import { ThemeProvider } from "@emotion/react";
 import { SubjectPageMainPane } from "../components/subject_page/main_pane";
 import { intl } from "../i18n/i18n";
-import { NamedTypedPlace, StatVarSpec } from "../shared/types";
+import { NamedTypedPlace } from "../shared/types";
 import theme from "../theme/theme";
 import { SubjectPageConfig } from "../types/subject_page_proto_types";
-import {
-  defaultDataCommonsClient,
-  defaultDataCommonsWebClient,
-} from "../utils/data_commons_client";
-import { TileSources } from "../utils/tile_utils";
-import {
-  isPlaceContainedInUsa,
-  pageMessages,
-  placeChartsApiResponsesToPageConfig,
-} from "./util";
+import { defaultDataCommonsWebClient } from "../utils/data_commons_client";
+import { PlaceOverview } from "./dev_place_overview";
+import { pageMessages, placeChartsApiResponsesToPageConfig } from "./util";
 
 /**
  * Component that renders the header section of a place page.
@@ -175,167 +165,6 @@ const PlaceTopicTabs = ({
 };
 
 /**
- * Component that displays a table of key demographic statistics for a place.
- *
- * Fetches data for population, median income, median age, unemployment rate,
- * and crime statistics using the Data Commons API. Displays the values in a
- * formatted table with units and dates.
- *
- * @param props.placeDcid The DCID of the place to show statistics for
- * @returns A table component showing key demographic statistics, or null if data not loaded
- */
-const PlaceOverviewTable = (props: {
-  placeDcid: string;
-}): React.JSX.Element => {
-  const { placeDcid } = props;
-  const [dataRows, setDataRows] = useState<DataRow[]>([]);
-  const containerRef = useRef(null);
-  // Fetch key demographic statistics for the place when it changes
-  useEffect(() => {
-    (async (): Promise<void> => {
-      const placeOverviewDataRows = await defaultDataCommonsClient.getDataRows({
-        entities: [placeDcid],
-        variables: [
-          "Count_Person",
-          "Median_Income_Person",
-          "Median_Age_Person",
-          "UnemploymentRate_Person",
-          "Count_CriminalActivities_CombinedCrime",
-        ],
-        perCapitaVariables: ["Count_CriminalActivities_CombinedCrime"],
-      });
-      setDataRows(placeOverviewDataRows);
-    })();
-  }, [placeDcid]);
-  if (!dataRows) {
-    return null;
-  }
-  const sourceUrls = new Set(
-    dataRows.map((dataRow) => {
-      return dataRow.variable.observation.metadata.provenanceUrl;
-    })
-  );
-  const statVarDcids = dataRows.map((dr) => {
-    return dr.variable.dcid;
-  });
-
-  const statVarSpecs: StatVarSpec[] = statVarDcids.map((dcid) => {
-    return {
-      statVar: dcid,
-      denom: "", // Initialize with an empty string or a default denominator if applicable
-      unit: "", // Initialize with an empty string or a default unit if applicable
-      scaling: 1, // Initialize with a default scaling factor
-      log: false, // Initialize with a default log value
-    };
-  });
-
-  return (
-    <table className="table" ref={containerRef}>
-      <thead>
-        <tr>
-          <th scope="col" colSpan={2}>
-            Key Demographics
-          </th>
-          <th scope="col"></th>
-        </tr>
-      </thead>
-      <tbody>
-        {dataRows.map((dataRow, index) => {
-          const unit = dataRow.variable.observation.metadata.unitDisplayName
-            ? dataRow.variable.observation.metadata.unitDisplayName
-            : "";
-          const formattedObservationValue =
-            dataRow.variable.observation.value.toLocaleString();
-          return (
-            <tr key={index}>
-              <td>{dataRow.variable.properties.name}</td>
-              <td>
-                {formattedObservationValue} {unit} (
-                {dataRow.variable.observation.date})
-              </td>
-              <td></td>
-            </tr>
-          );
-        })}
-        {dataRows && (
-          <tr>
-            <td>
-              <div className="chart-container">
-                <TileSources
-                  containerRef={containerRef}
-                  sources={sourceUrls}
-                  statVarSpecs={statVarSpecs}
-                />
-              </div>
-            </td>
-            <td></td>
-            <td></td>
-          </tr>
-        )}
-      </tbody>
-    </table>
-  );
-};
-
-/**
- * Displays an overview of a place including its name, summary, map and key statistics.
- *
- * @param props.place The place object containing name and dcid
- * @param props.placeSummary A text summary describing the place
- * @returns A component with the place overview including icon, name, summary, map and statistics table
- */
-const PlaceOverview = (props: {
-  place: NamedTypedPlace;
-  placeSummary: string;
-  parentPlaces: NamedTypedPlace[];
-}): React.JSX.Element => {
-  const { place, placeSummary, parentPlaces } = props;
-  const isInUsa = isPlaceContainedInUsa(
-    parentPlaces.map((place) => place.dcid)
-  );
-  const theme = useTheme();
-  return (
-    <div className="place-overview">
-      <div
-        css={css`
-          ${theme.typography.text.md}
-          align-items: center;
-          display: flex;
-          font-weight: 500;
-          gap: ${theme.spacing.sm}px;
-          margin-bottom: ${theme.spacing.md}px;
-          & > svg {
-            height: 1.5rem;
-          }
-        `}
-      >
-        <LocationCity />
-        <span>{intl.formatMessage(pageMessages.SummaryOverview)}</span>
-      </div>
-      <div
-        css={css`
-          ${theme.typography.text.sm}
-          margin-bottom: ${theme.spacing.md}px;
-        `}
-      >
-        {placeSummary}
-      </div>
-      <div className="row place-map">
-        {isInUsa && (
-          <div className="col-md-3">
-            <GoogleMap dcid={place.dcid}></GoogleMap>
-          </div>
-        )}
-        <div className="col-md-9">
-          {!isInUsa && <br></br>}
-          <PlaceOverviewTable placeDcid={place.dcid} />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-/**
  * Component that displays a list of child places for a given place.
  *
  * @param props.place The parent place containing name and dcid
@@ -352,7 +181,7 @@ const RelatedPlaces = (props: {
     return null;
   }
 
-  const NUM_PLACES = 15;
+  const NUM_PLACES = 100;
   const showToggle = childPlaces.length > NUM_PLACES;
   const truncatedPlaces = childPlaces.slice(0, NUM_PLACES);
   const numPlacesCollapsed = childPlaces.length - NUM_PLACES;
